requirements.txt
scikit-learn
pandas
matplotlib
python-dotenv
pydot
jupyterlab
keras-tuner
jupyter
boto3
langchain
langchain-community
langchain-aws
langchain_openai
openai
neo4j
konlpy
statsmodels